Updating with current collection service code

+++ app/src/main/java/com/eveningoutpost/dexdrip/Models/TransmitterData.java
@@ -55,7 +55,7 @@ public class TransmitterData extends Model {
             transmitterData.filtered_data =txData.getInt(6);
             Log.i(TAG, "Filtered Data: " + transmitterData.filtered_data);
             transmitterData.sensor_battery_level = txData.getShort(10);
-            transmitterData.wixel_battery_level = txData.getInt(11);
+            transmitterData.wixel_battery_level = txData.getShort(11);
             //transmitterData.wixel_battery_level = txData.get(11);
             Log.i(TAG, "Wix Batt: " + transmitterData.wixel_battery_level);
             transmitterData.timestamp = new Date().getTime();
